construct package directory
