+++ src/kanbanwave/components/CardDraggable.tsx
