+++ src/kanbanwave/components/NewCard.tsx
@@ -1,13 +1,14 @@
-import { ChangeEvent, useCallback, useState } from 'react';
-import { Button, IconButton, TextArea } from 'app/components';
 import clsx from 'clsx';
+import { ChangeEvent, useCallback, useEffect, useState } from 'react';
+import { Button, IconButton, TextArea } from 'app/components';
 
-type NewBoardProps = {
+type NewCardProps = {
   className?: string;
+  cardsLength?: number;
   onAdd?: (title: string) => void;
 };
 
-const NewBoard = ({ className, onAdd }: NewBoardProps) => {
+const NewCard = ({ className, cardsLength, onAdd }: NewCardProps) => {
   const [isInputVisible, setIsInputVisible] = useState(false);
   const [title, setTitle] = useState('');
 
@@ -15,32 +16,43 @@ const NewBoard = ({ className, onAdd }: NewBoardProps) => {
     setTitle(e.target.value);
   }, []);
 
-  const handleAddClick = useCallback(() => {
-    if (title.trim() !== '') {
-      onAdd?.(title);
-      setIsInputVisible(false);
-      setTitle('');
-    }
-  }, [title, onAdd]);
+  const handleAddClick = useCallback(
+    (e: React.MouseEvent<HTMLButtonElement>) => {
+      e.preventDefault();
+      if (title.trim() !== '') {
+        onAdd?.(title);
+        setTitle('');
+      }
+    },
+    [title, onAdd]
+  );
 
   const handleCancelClick = () => {
-    setIsInputVisible(false);
     setTitle('');
+    setIsInputVisible(false);
   };
 
+  useEffect(() => {
+    if (cardsLength === 0) {
+      setIsInputVisible(true);
+    } else {
+      setIsInputVisible(false);
+    }
+  }, [cardsLength]);
+
   return (
-    <div className="h-[272px] rounded-lg w-[23%] bg-zinc-50 shrink-0">
+    <div className="flex flex-col w-full rounded-lg max-w-64 bg-zinc-50 shrink-0 h-fit">
       {isInputVisible ? (
         <div className={clsx('p-2', className)}>
           <TextArea
-            placeholder={`Enter a board title`}
+            placeholder={`Enter a card title`}
             value={title}
             onChange={handleChange}
             className="w-full py-1 leading-8 min-h-11 "
           />
           <div className="flex items-start mt-1">
             <Button type="button" className="mr-2" onClick={handleAddClick}>
-              Add board
+              Add card
             </Button>
             <IconButton
               name="close"
@@ -54,14 +66,14 @@ const NewBoard = ({ className, onAdd }: NewBoardProps) => {
       ) : (
         <IconButton
           name="add"
-          aria-label={`add a board`}
+          aria-label={`add a card`}
           onClick={() => setIsInputVisible(true)}
-          className={clsx('w-full h-full')}>
-          <span className="text-sm">Add another board</span>
+          className={clsx('h-11', 'mt-1')}>
+          <span className="text-sm">Add a card</span>
         </IconButton>
       )}
     </div>
   );
 };
 
-export default NewBoard;
+export default NewCard;
