+++ src/kanbanwave/components/StrictModeDroppable.tsx
